Update A4.R
Overall good job. Exercises 4-5 are very good. You are commenting perfectly. 

However, the coding and naming are very heaving and hence hard to follow.. Because the assignment is long, these things carry over. There are too many un-necessary functions.  It is okay to write a function to do a basic operation but when you start to write function to sum two elements.. That is too much.